requirements.txt
requests>=2.31.0
PyGithub>=1.59.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
ruff>=0.1.0
coverage>=7.0.0
